pyproject.toml
[tool.poetry]
name = "floodgate-update-slack"
version = "0.1.0"
description = ""
authors = ["select766 <select766@outlook.jp>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
